requirements.txt
asgiref==3.6.0
./diploma-frontend-0.2.tar.gz
Django==4.1.7
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2023.2
six==1.16.0
sqlparse==0.4.3
